Pull BasicTabs function from App.js to BasicTabs.js
