+++ .idea/.gitignore
@@ -3,3 +3,5 @@
 /workspace.xml
 # Editor-based HTTP Client requests
 /httpRequests/
+/package-lock.json
+/package.json
